Separated loading component to be reused on suspense in lazy imported modules | Improved menu and layout organization | Moved menu components to home feature
