+++ react/src/components/loading.component.tsx
@@ -1,15 +1,13 @@
 import LoadingSpinner from '@assets/looties/loading-spinner/loading-spinner.lootie';
-import { useIsFetching, useIsMutating } from '@tanstack/react-query';
+import { MENU_NAVBAR_HEIGHT } from '@constants/styles.constants';
 
 export default function Loading() {
-  const isFetching = useIsFetching();
-  const isMutating = useIsMutating();
-
   return (
-    (isFetching || isMutating) ?
-      <div className="h-screen w-screen bg-black opacity-25 absolute z-50 flex justify-center items-center">
-        <LoadingSpinner style={{ height: '10rem' }} />
-      </div> :
-      <div />
+    <div
+      className="w-screen bg-black opacity-25 absolute z-50 flex justify-center items-center"
+      style={{ height: `calc(100vh - ${MENU_NAVBAR_HEIGHT})` }}
+    >
+      <LoadingSpinner style={{ height: '10rem' }} />
+    </div>
   );
 }
