+++ react/src/constants/styles.constants.ts
@@ -1,0 +1,1 @@
+export const MENU_NAVBAR_HEIGHT = '4rem';
