+++ react/src/features/home/components/menu-layout.component.tsx
@@ -1,0 +1,11 @@
+import Layout from './layout.component';
+import Menu from './menu.component';
+
+export default function MenuLayout() {
+  return (
+    <>
+      <Menu />
+      <Layout />
+    </>
+  );
+}
