--- react/src/features/menu/components/index.ts
@@ -1,5 +1,0 @@
-import Menu from './menu.component';
-
-export {
-  Menu
-};
