--- react/src/features/menu/index.ts
@@ -1,1 +1,0 @@
-export * from './components';
